# server/swagger.yaml
openapi: 3.0.0
info:
  title: CadenceRT
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /my/advanced/collectives/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Collective
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Collective
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collective"
    post:
      summary: CRUD ACTION update
      tags:
        - Collective
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collective"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Collective"
  /my/advanced/collectives/findByr_members/{key}:
    get:
      summary: CRUD ACTION findByr_members
      tags:
        - Collective
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collective"
  /my/advanced/collectives/findByr_owner/{key}:
    get:
      summary: CRUD ACTION findByr_owner
      tags:
        - Collective
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collective"
    
  #-----  Filesystem  -----
...
  /my/filesystems/{id}:
    delete:
      summary: Delete a file. Admins only.
      tags:
        - Filesystem
      parameters:
        - name: id
          in: path
          description: Id Filesystem path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: Download or receive a file from the database.
      tags:
        - Filesystem
      parameters:
        - name: id
          in: path
          description: The path to file
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filesystem"
    post:
      summary: CRUD ACTION update
      tags:
        - Filesystem
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filesystem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Filesystem"
  /my/filesystems/findByfilename/{key}:
    get:
      summary: Search for a file by filename
      tags:
        - Filesystem
      parameters:
        - name: key
          in: path
          description: filename of the resource filename to search
          required: true
          schema:
            type: string
        - name: owner
          in: path
          description: the file owners reational Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Filesystem"
  /my/filesystems/findByr_owner/{key}:
    get:
      summary: CRUD ACTION findByr_owner
      tags:
        - Filesystem
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Filesystem"
  /my/filesystems/{id}/getr_owner:
    get:
      summary: List all the files belonging to a user.
      tags:
        - Filesystem
      parameters:
        - name: id
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Geocode  -----
...
  /user/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
  #-----  User  -----
  /Users/{id}/changePassword:
    post:
      summary: Change password by admin API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordAdmin:
                  type: string
      responses:
        '200':
          description: OK
  /Users/{id}:
    post:
      summary: Update User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    delete:
      summary: Delete User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Collective:
      properties:
        _id:
          type: string
        _created:
          type: integer
        _expiration:
          type: integer
        _modified:
          type: integer
        class:
          type: string
        description:
          type: string
        handle:
          type: string
        locked:
          type: boolean
        metadata:
          type: string
        status:
          type: string
        subclass:
          type: string
        token:
          type: string
      required:
        -  _id
        - _created
        - _modified
        - class
        - handle
        - locked
        - status
    Filesystem:
      properties:
        _id:
          type: string
        _created:
          type: integer
        _modified:
          type: integer
        _status:
          type: string
        basepath:
          type: string
        blob:
          type: number
        filename:
          type: string
        filesize:
          type: number
        locked:
          type: string
        mime:
          type: number
        permission:
          type: integer
        r_user:
          type: number
      required:
        -  _id
        - _created
        - _status
        - basepath
        - filename
        - locked
        - permission
...
    User:
      properties:
        _id:
          type: string
        _created:
          type: integer
        _expires:
          type: integer
        _lastactive:
          type: integer
        address_1:
          type: string
        address_2:
          type: string
        address_3:
          type: string
        address_city:
          type: string
        address_country:
          type: string
        address_postalcode:
          type: string
        avatar:
          type: integer
        id_number:
          type: string
        locale:
          type: string
        mail_primary:
          type: string
        msisdn:
          type: string
        name_first:
          type: string
        name_last:
          type: string
        name_title:
          type: string
        nationality:
          type: string
        opt_ins:
          type: string
        opt_outs:
          type: string
        password:
          type: string
        r_group:
          type: integer
        r_parent:
          type: integer
        r_policy:
          type: integer
        r_role:
          type: string
        status:
          type: string
        super_class:
          type: object
        tax_number:
          type: string
        username:
          type: string
      required:
        -  _id
        - mail_primary
        - msisdn
        - password
        - r_role
        - status
        - super_class
        - username